backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
python-jose[cryptography]
websockets
pydantic[email]
python-multipart
werkzeug>=2.3.0
